toolbox/requirements.txt
# TouhouRL Toolbox 依赖包

# 屏幕捕获
mss>=9.0.0

# 键盘控制
pyautogui>=0.9.54

# 图像处理
Pillow>=10.0.0
opencv-python>=4.8.0

# 数值计算
numpy>=1.24.0
